reorganize exchange crate
